Add suricata-update to build
This change adds a pyinstaller-frozen suricata-update to our suricata
build on all three platforms.

It also adds a "suricataupdater" executable (shell script on unix-y
platforms, go binary on Windows) following the same strategy that we
took for the suricatarunner and zeekrunner.

As part of this work I had to fork suricata-update to fix some Windows
issues. This fork is used here pending the fixes being upstreamed.