Make suricata helpers userdir-aware
This commit changes the suricata helpers to take into consideration a
`BRIM_SURICATA_USER_DIR` environment variable. If present, the helpers
use that dir to write  suricata config and rules.

If the environment variable is unset, the helpers continue to use the
existing installation directory as previously.

This is motivated by Linux system wide installs where the install dir
is non-writable.